progress bar refactor
